Выполнил задание #12. Личный проект: капитан ООП
